core: switch br_angle from br_fixed_luf to br_scalar
Angles being 0.16 is limiting, especially when rotating things at higher
framerates.

- Redefines br_angle to be br_scalar instead of br_fixed_luf.
- Update support macros to
- Change the signatures of BrFixed*() to take a br_fixed_luf instead,
  keeping API compatibility.
- Add BrFixedLUFToScalar() and BrScalarToFixedLUF() helper macros
- Add new datafile chunk types for each structure that contains an
  angle, while keeping compatibility for loading fixed angles.

+++ core/fw/tokenval.c
@@ -2093,7 +2093,7 @@ br_value BR_RESIDENT_ENTRY BrTokenValueVaArg(br_token token, va_list *ap)
             return (br_value){.f = (float)va_arg(*ap, double)};
         /* NB: missing BRT_SCALAR */
         case BRT_ANGLE:
-            return (br_value){.a = (br_angle)va_arg(*ap, int)};
+            return (br_value){.a = (br_angle)va_arg(*ap, double)};
         case BRT_COLOUR_RGB:
             return (br_value){.rgb = va_arg(*ap, br_colour)};
         case BRT_OBJECT:
